Fixed 2 initial components

+++ src/graph/geometric.h
@@ -22,10 +22,10 @@ public:
 
 private:
 
-    std::vector<int> generatePosition() const;
+    std::vector<float> generatePosition() const;
 
     // pre: v1.size() == v2.size()
-    static float calculateDistance(std::vector<int> const& v1, std::vector<int> const& v2);
+    static float calculateDistance(std::vector<float> const& v1, std::vector<float> const& v2);
 
     unsigned int n, dimension;
     float radius;
